Problem: unable to capture errors
This prevents us from handling errors and doing something if they occur.

Solution: introduce TRY word that safely evaluates the closure and pushes an
error closure or an empty closure onto the stack.

While at it, I discovered that `error_program!` macro over-encoded closures
(because the errors are supposed to go on to the stack), so it led to something
like this:

```
[DUP] TRY UNWRAP UNWRAP
0x456D70747920737461636B 0x 0x4
```

(while it should be just one UNWRAP, of course)

So I fixed that and `assert_error!` macro as well.